WLIAS-0020
Added Vector2f and Vector3f classes, these support vectors of size 2 and 3 respectively, usually used for coordinates. Added a Mesh class which initializes a mesh with vertices and indices. Added a Vertex2D class which receives a Vector2F class and holds its data as a position. Finally added  a Renderer class which reads in a Mesh and draws it to the screen.