WLIAS-0025
Added Matrix4f Transformations and a GameObject.java. We can now render GameObjects with added information, position, scale and rotation. A GameObject also hold a Mesh. Our Renderer.java now can receive a GameObject and read it's data to draw to the screen, it also applies uniforms to dynamically change this data in the shader. Unit Tests have been added to test all f this functionality