WLIAS-0041 Added Initial Pause Menu
Added a pause menu that is triggered upon pressing the escape key. The menu freezes the camera and halts the game. From the menu you can wither resume by clicking the resume button or pressing the escape key again. The main menu button can also be clicked in ordered to destroy the world and return to the main menu.